Add source db to product service

+++ services/productos.js
@@ -1,9 +1,10 @@
 const fs = require('fs');
-const products = [];
 
+const { config } = require('../config/index');
 const pathFile = process.env.PATH_PRODUCTOS;
-
+let products = [];
 const { ProductosModel } = require('../models/Productos');
+
 
 const connectDB = async () => {
     try {
@@ -17,78 +18,105 @@ const connectDB = async () => {
 }
 
 const saveProduct = async (product) => {
-    //await connectDB();
+
+    let id;
+
+    if (config.database === 'FILESYSTEM') {
+        
+        products.push(product);
+        await fs.promises.writeFile(pathFile, JSON.stringify(products, null, 2)); 
+        id = product.id; 
     
-    /*
-    products.push(product);
-    await fs.promises.writeFile(pathFile, JSON.stringify(products, null, 2));  
-    return product.id;
-    */
-   let id;
-   try {
-        const prd = await ProductosModel.create(product);
-        id = prd._id;
-   } catch (err) {
-        console.log(err);
-   }
+    } else if (config.database === 'MONG') {
+
+        try {
+            const prd = await ProductosModel.create(product);
+            id = prd._id;
+        } catch (err) {
+            console.log(err);
+        }
+    
+    }
+
 
    return id;
 }
 
 const findProductById = async (id) => {
-    /*
-    const product = products.find(prd => prd.id === id && (!prd.borrado) ); //be8c11dd-3bc3-43b7-91c2-cf43a3c15757
-    return product;
-    */
+
     let prd, countPrd;
-    try {
+
+    if (config.database === 'FILESYSTEM') {
+
         if (id !== undefined ) {
-            prd = await ProductosModel.findById(id); 
+            prd = products.find(prd => prd.id === id);
         } else {
-            //prd = await ProductosModel.find({ borrado: false });
-            prd = await ProductosModel.find({});
-            countPrd = await ProductosModel.countDocuments({});
+            countPrd = products.length;
             prd = {
                 total: countPrd,
-                products: prd
+                prd: products
             }
         }
         
-    } catch (err) {
-       console.log(err);
+
+    } else if (config.database === 'MONG') {
+
+        try {
+            if (id !== undefined ) {
+                prd = await ProductosModel.findById(id); 
+            } else {
+                //prd = await ProductosModel.find({ borrado: false });
+                prd = await ProductosModel.find({});
+                countPrd = await ProductosModel.countDocuments({});
+                prd = {
+                    total: countPrd,
+                    products: prd
+                }
+            }
+            
+        } catch (err) {
+           console.log(err);
+        }
     }
+
 
     return prd;
 
 }
 
 const updateProductById = async (id, prd) => {
-    /*
-    const indexProduct = products.findIndex(prd => prd.id === id && (!prd.borrado));
-    let product;
+ 
+    let prdUpd, product, indexProduct;
 
-    if (indexProduct !== -1) {
-        //Actualizo el producto.
-       products[indexProduct].nombre = prd.nombre !== undefined ? prd.nombre : products[indexProduct].nombre ;
-       products[indexProduct].descripcion = prd.descripcion !== undefined ? prd.descripcion : products[indexProduct].descripcion ;
-       products[indexProduct].codigo = prd.codigo !== undefined ? prd.codigo : products[indexProduct].codigo ;
-       products[indexProduct].foto = prd.foto !== undefined ? prd.foto : products[indexProduct].foto ;
-       products[indexProduct].precio = prd.precio !== undefined ? prd.precio : products[indexProduct].precio ;
-       products[indexProduct].stock = prd.stock !== undefined ? prd.stock : products[indexProduct].stock ;
-       product = products[indexProduct];
-       await fs.promises.writeFile(pathFile, JSON.stringify(products, null, 2));
+    if (config.database === 'FILESYSTEM') {
+
+        indexProduct = products.findIndex(prd => prd.id === id);
+
+        if (indexProduct !== -1) {
+            //Actualizo el producto.
+           products[indexProduct].nombre = prd.nombre !== undefined ? prd.nombre : products[indexProduct].nombre ;
+           products[indexProduct].descripcion = prd.descripcion !== undefined ? prd.descripcion : products[indexProduct].descripcion ;
+           products[indexProduct].codigo = prd.codigo !== undefined ? prd.codigo : products[indexProduct].codigo ;
+           products[indexProduct].foto = prd.foto !== undefined ? prd.foto : products[indexProduct].foto ;
+           products[indexProduct].precio = prd.precio !== undefined ? prd.precio : products[indexProduct].precio ;
+           products[indexProduct].stock = prd.stock !== undefined ? prd.stock : products[indexProduct].stock ;
+           product = products[indexProduct];
+           await fs.promises.writeFile(pathFile, JSON.stringify(products, null, 2));
+        }
+
+        prdUpd = product;
+
+    } else if (config.database === 'MONG') {
+        try {
+
+            prdUpd = await ProductosModel.findByIdAndUpdate(id, prd, { new: true});
+    
+        } catch (err) {
+            console.log(err);
+        }
     }
-    return product;
-    */
-    let prdUpd;
 
-    try {
-
-        prdUpd = await ProductosModel.findByIdAndUpdate(id, prd, { new: true});
-
-    } catch (err) {
-        console.log(err);
-    }
+    
 
     return prdUpd;
 }
@@ -102,13 +130,27 @@ const deleteProductById = async (id) => {
     }
     return products[indexProduct];
     */
-   let prd;
+    let prd, indexProduct, productsAux;
 
-   try {
-        prd = await ProductosModel.findByIdAndDelete(id);
-   } catch (err) {
-       console.log(err);
-   }
+    if (config.database === 'FILESYSTEM') {
+        
+        indexProduct = products.findIndex(prd => prd.id === id );
+
+        if (indexProduct !== -1) {
+            prd = products[indexProduct]; 
+            //products[indexProduct].borrado = true;
+            products = products.filter(prd => prd.id !== id );
+            
+            await fs.promises.writeFile(pathFile, JSON.stringify(products, null, 2));
+        }
+
+    } else if (config.database === 'MONG') {
+        try {
+            prd = await ProductosModel.findByIdAndDelete(id);
+        } catch (err) {
+            console.log(err);
+        }
+    }
 
    return prd;
 }
